requirements.txt
pymupdf==1.24.12
fastapi==0.105.0
uvicorn==0.24.0
python-multipart==0.0.6

# Additional libraries used by the application
pillow
pytesseract
pydantic
